backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-jose
python-dotenv
email-validator
neo4j
httpx
passlib[bcrypt]==1.7.4
bcrypt==3.2.2